Adrressed some PR comments: - Removed a plugin and merged the code with with the newly named TimelineTransformationPlugin - Added french Translation of Intake Officer - Modified tetsts a bit - Removed a field under reminder entity
