+++ src/Compliance.Plugins.Tests/Compliance.Plugins.Tests.csproj
@@ -48,8 +48,8 @@
     <Reference Include="FakeXrmEasy, Version=1.57.0.0, Culture=neutral, PublicKeyToken=dd4e58c4492635a6, processorArchitecture=MSIL">
       <HintPath>..\packages\FakeXrmEasy.9.1.57.0\lib\net452\FakeXrmEasy.dll</HintPath>
     </Reference>
-    <Reference Include="FluentAssertions, Version=5.10.0.0, Culture=neutral, PublicKeyToken=33f2691a05b67b6a, processorArchitecture=MSIL">
-      <HintPath>..\packages\FluentAssertions.5.10.0\lib\net47\FluentAssertions.dll</HintPath>
+    <Reference Include="FluentAssertions, Version=5.10.3.0, Culture=neutral, PublicKeyToken=33f2691a05b67b6a, processorArchitecture=MSIL">
+      <HintPath>..\packages\FluentAssertions.5.10.3\lib\net47\FluentAssertions.dll</HintPath>
     </Reference>
     <Reference Include="Microsoft.Crm.Sdk.Proxy, Version=9.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <HintPath>..\packages\Microsoft.CrmSdk.CoreAssemblies.9.0.2.22\lib\net462\Microsoft.Crm.Sdk.Proxy.dll</HintPath>
@@ -80,6 +80,7 @@
     <Reference Include="System" />
     <Reference Include="System.Activities" />
     <Reference Include="System.Activities.Presentation" />
+    <Reference Include="System.Configuration" />
     <Reference Include="System.Core" />
     <Reference Include="System.Data" />
     <Reference Include="System.Data.DataSetExtensions" />
@@ -147,13 +148,12 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AnnotationAuthorPluginTests.cs" />
+    <Compile Include="TimelineTransformationsTests.cs" />
     <Compile Include="CompliancePrefixPluginTests.cs" />
     <Compile Include="QueueMembershipPluginTests.cs" />
     <Compile Include="MultiLanguagePluginTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ChangeEventPluginTests.cs" />
-    <Compile Include="RemoveTimelineCardStatusPluginTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
