+++ src/Compliance.Plugins.Tests/packages.config
@@ -2,7 +2,7 @@
 <packages>
   <package id="FakeItEasy" version="3.2.0" targetFramework="net471" />
   <package id="FakeXrmEasy.9" version="1.57.0" targetFramework="net471" />
-  <package id="FluentAssertions" version="5.10.0" targetFramework="net471" />
+  <package id="FluentAssertions" version="5.10.3" targetFramework="net471" />
   <package id="Microsoft.CrmSdk.CoreAssemblies" version="9.0.2.22" targetFramework="net471" />
   <package id="Microsoft.CrmSdk.Deployment" version="9.0.2.22" targetFramework="net471" />
   <package id="Microsoft.CrmSdk.Workflow" version="9.0.2.22" targetFramework="net471" />
